Use _AsyncPeriodic for ClientSession too
This involved moving it to bokeh.util.tornado and
making it support non-async callbacks as well.
